backend/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.0
slack-sdk>=3.27.1
python-multipart>=0.0.9
pytest>=7.4.3
pytest-asyncio>=0.21.1
pydantic>=2.9.0
web3>=6.11.0
eth-account>=0.10.0
cryptography>=41.0.0
# Security & Authentication
PyJWT>=2.8.1
passlib>=1.7.4
python-jose>=3.3.0
pyotp>=2.9.0
qrcode>=7.4.2
# Email validation
email-validator>=2.1.0
# SSL/TLS
pyopenssl>=24.0.0
# Additional security
numpy>=1.24.0
scikit-learn>=1.3.0